claude: refactor(kaspa): move validator module from libs/kaspa to rust/main
- Move validator module from libs/kaspa/lib/validator to rust/main/chains/dymension-kaspa/src/kas_validator
- Update imports: crate::error -> crate::kas_validator::error, corelib -> dym_kas_core, bridge -> dym_kas_bridge, api_rs -> dym_kas_api, secp256k1 -> kaspa_bip32::secp256k1
- Update dymension-kaspa lib.rs to expose kas_validator module
- Update validator_server.rs to use local kas_validator imports
- Remove dym-kas-validator dep from dymension-kaspa and hyperlane-base Cargo.toml
- Update libs/kaspa kms crate: remove validator dep, define KaspaSecpKeypair locally
- Update libs/kaspa tooling: depend on dymension-kaspa for signer module, add ethers feature to hyperlane-core
- Remove validator from libs/kaspa workspace members
