+++ dymension/libs/kaspa/lib/kms/Cargo.toml
@@ -12,6 +12,3 @@ eyre = { workspace = true }
 tracing = { workspace = true }
 serde_json = { workspace = true }
 secp256k1 = { workspace = true }
-
-# Kaspa validator keypair
-dym-kas-validator = { path = "../validator", package = "validator" }
