--- dymension/libs/kaspa/lib/validator/Cargo.toml
@@ -1,40 +1,0 @@
-[package]
-name = "validator"
-version = "0.2.0"
-edition = "2021"
-
-[dependencies]
-corelib = { path = "../core", package = "core" }
-bridge = { path = "../bridge" }
-api-rs = { path = "../api", package = "openapi" }
-hardcode = { path = "../hardcode", package = "hardcode" }
-dymension-kaspa-hl-constants = { path = "../../../../../rust/main/chains/dymension-kaspa-hl-constants" }
-kaspa-addresses = { workspace = true }
-kaspa-consensus-core = { workspace = true }
-kaspa-core = { workspace = true }
-kaspa-grpc-client = { workspace = true }
-kaspa-wallet-core = { workspace = true }
-kaspa-wallet-pskt = { workspace = true }
-kaspa-txscript = { workspace = true }
-kaspa-rpc-core = { workspace = true }
-kaspa-wrpc-client = { workspace = true }
-kaspa-hashes = { workspace = true }
-secp256k1 = { workspace = true }
-workflow-core = { workspace = true }
-tokio = { workspace = true, features = ["full"] }
-tracing = { workspace = true }
-tracing-error = { workspace = true }
-tracing-futures = { workspace = true }
-log = { workspace = true }
-axum = { workspace = true }
-hyperlane-core = { path = "../../../../../rust/main/hyperlane-core" , features = ["agent", "async", "test-utils"] }
-hyperlane-warp-route = { path = "../../../../../rust/main/applications/hyperlane-warp-route" }
-hyperlane-cosmos = { path = "../../../../../rust/main/chains/hyperlane-cosmos" }
-hyperlane-cosmos-rs = { workspace = true }
-eyre = { workspace = true }
-serde = { version = "^1.0", features = ["derive"] }
-anyhow = "1.0" # TODO: prefer eyre
-ethers = "2.0"
-serde_json = { workspace = true }
-thiserror = "1.0"
-hex = { workspace = true }
