--- dymension/libs/kaspa/lib/validator/README.md
@@ -1,3 +1,0 @@
-## What?
-
-Contains logic and types used by validator but not relayer.
