+++ dymension/libs/kaspa/tooling/Cargo.toml
@@ -8,7 +8,7 @@ kaspa-core = { workspace = true }
 corelib = { path = "../lib/core", package = "core" }
 bridge = { path = "../lib/bridge" }
 hardcode = { path = "../lib/hardcode" }
-validator = { path = "../lib/validator" }
+dymension-kaspa = { path = "../../../../rust/main/chains/dymension-kaspa" }
 secp256k1 = { workspace = true }
 serde = { version = "^1.0", features = ["derive"] }
 serde_json = { workspace = true }
@@ -24,7 +24,7 @@ kaspa-consensus-core = { workspace = true }
 kaspa-addresses = { workspace = true }
 workflow-core = { workspace = true }
 uuid = { version = "1.11", features = ["v4"] }
-hyperlane-core = { path = "../../../../rust/main/hyperlane-core" }
+hyperlane-core = { path = "../../../../rust/main/hyperlane-core", features = ["ethers"] }
 hyperlane-cosmos = { path = "../../../../rust/main/chains/hyperlane-cosmos" }
 hyperlane-metric = { path = "../../../../rust/main/hyperlane-metric" }
 rand = { version = "0.9" }
