+++ dymension/libs/kaspa/tooling/src/x/relayer.rs
@@ -1,4 +1,4 @@
-use validator::signer::{get_ethereum_style_signer, EthereumStyleSigner};
+use dymension_kaspa::kas_validator::signer::{get_ethereum_style_signer, EthereumStyleSigner};
 
 pub fn create_relayer() -> EthereumStyleSigner {
     get_ethereum_style_signer().unwrap()
