+++ dymension/libs/kaspa/tooling/src/x/validator.rs
@@ -2,7 +2,7 @@ use corelib::escrow::generate_escrow_priv_key;
 use secp256k1::PublicKey;
 use serde::Serialize;
 use std::fs;
-use validator::signer::get_ethereum_style_signer;
+use dymension_kaspa::kas_validator::signer::get_ethereum_style_signer;
 
 use super::args::{ValidatorAwsArgs, ValidatorLocalArgs};
