+++ rust/main/chains/dymension-kaspa/src/kas_validator/withdraw_test.rs
@@ -1,12 +1,12 @@
 // TODO: some signarures have changed and the test fails to compile
 // #[cfg(test)]
 // mod relayer_validator_flow_tests {
-//     use crate::error::ValidationError;
+//     use crate::kas_validator::error::ValidationError;
 //     use crate::withdraw::validate_pskts;
-//     use corelib::escrow::{Escrow, EscrowPublic};
-//     use bridge::payload::{MessageID, MessageIDs};
-//     use bridge::util::get_recipient_script_pubkey_address;
-//     use bridge::withdraw::WithdrawFXG;
+//     use dym_kas_core::escrow::{Escrow, EscrowPublic};
+//     use dym_kas_bridge::payload::{MessageID, MessageIDs};
+//     use dym_kas_bridge::util::get_recipient_script_pubkey_address;
+//     use dym_kas_bridge::withdraw::WithdrawFXG;
 //     use eyre::{eyre, Result}; // Use eyre's Result for convenient error handling
 //     use hyperlane_core::{Encode, HyperlaneMessage, H256, U256};
 //     use hyperlane_warp_route::TokenMessage;
